Update patch set 3
Patch Set 3:

These were intentionally not macros to avoid recalculating the arguments multiple times. In the current use, this just looks wasteful, not unsafe, but if someone were to use, say, an increment or decrement operator inside a call it'd become wrong.

Unless there's a major motivation for this change I'd rather revert it.

Patch-set: 3
CC: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>