Update patch set 2
Patch Set 2:

Looks like this didn't help at all. We could maybe add __attribute(noinline), but I think that's not a good way to solve this.

I hope to try and add this attribute to vector itself to fix such cases as this. Feel free to follow along here:
https://bugs.chromium.org/p/chromium/issues/detail?id=738155

Patch-set: 2