Update patch set 6
Patch Set 6:

The other uses of the flag were actually not needed so I removed them in this CL as well. Then I just check if the flag is defined in the one file where it is used, and define it if not.

PTAL =]

Patch-set: 6
Reviewer: Gerrit User 1167239 <1167239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0