[unicode] Return (the correct) errors for overlong / surrogate sequences.
This fix is two-fold:

1) Incremental UTF-8 decoding: Unify incorrect UTF-8 handling between V8 and
Blink.

Incremental UTF-8 decoding used to allow some overlong sequences which Blink
treated as errors. This caused the decoder and the Blink UTF-8 decoder to
produce a different number of bytes, resulting in random failures when scripts
were streamed (especially, this was detected by the skipping inner functions
feature which adds CHECKs against expected function positions).

2) Non-incremental UTF-8 decoding: return the correct amount of invalid characters.

According to the encoding spec ( https://encoding.spec.whatwg.org/#utf-8-decoder
), the first byte of an overlong / surrogate sequence generates an invalid
character, and the rest of the bytes are not processed (i.e., pushed back to the
byte stream). When they're processed, they will look like lonely continuation
bytes, and will generate an invalid character each.

As a result, an overlong 4-byte sequence should generate 4 invalid characters
(not 1).

This is a potentially breaking change, since the (non-incremental) UTF-8
decoding is exposed via the API (String::NewFromUtf8). The behavioral difference
happens when the client is passing in invalid UTF-8 (containing overlong /
surrogate sequences).

BUG:chromium:765608, chromium:758236
Bug:
Change-Id: Ib029f6a8e87186794b092e4e8af32d01cee3ada0