Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Ali, does this contain a fix for the "top_on_previous_step" dcheck failure?

Not yet. I am working on that. The CL is not ready for review just yet.

It is possible that my workflow is not the best. I normally create the reland CL and then patch it in locally, work on the fix, and the upload and ask for review.

Is that not the best way to do this?

Secondly, is there an easy for me to run the failing GPU test against my fix? I don't have the environment to be able to run the Chrome, and specifically the GPU tests locally.

Patch-set: 1
Reviewer: Gerrit User 1148031 <1148031@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>