Update patch set 3
Patch Set 3:

Yes, sure, I wait for this, thank you for checking it! Is it correct that we do not have a bot for this configuration on the waterfall and it's impossible to build on a an amd64 machine?

Patch-set: 3
Reviewer: Gerrit User 1180419 <1180419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>