Update patch set 6
Patch Set 6:

> The chunking strategy seems good to me. I can imagine fancier things like chunking up in a binary tree sort of style, which could reduce the number of divisions, but I imagine that would only improve performance on really really big numbers, not necessarily something we need to be optimizing for at this stage.

I don't follow... I think if we want to end up with n chunks (of size <= maximum digit), then we need n-1 divisions. If we organize the sequence of divisions differently, they will operate on smaller dividends (so they might be faster) and we could (theoretically) parallelize processing of branches of the tree, but the price we'd pay is keeping more temporary results in memory, and doing BigInt multiplications to construct the bigger divisors. So I'm not convinced it would be an improvement.

Patch-set: 6
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0