[wasm] Add negative guard regions
Although Wasm memory indices are all unsigned, they sometimes get assembled
as 32-bit signed immediates. Values in the top half of the Wasm memory space
will then get sign extended, causing Wasm to access in front of its memory
buffer.

Usually this region is not mapped anyway, so faults still happen as they are
supposed to. This change protects this region with guard pages so we are
guaranteed to always fault when this happens.

See https://crrev.com/c/484747 for a previous attempt. The code is much cleaner
now due to having separated the allocation from the memory in the meantime
(https://crrev.com/c/505715).

Bug:
Change-Id: Ia034b49c00e63c3534429b5088f6175a862d3e27