Update patch set 7
Patch Set 7:

> Patch Set 4:
> 
> Please add regression tests that exercise each of the changes you've made to the ICU calls.

I was about to add a test for Runtime_GetLanguageTagVariants only to find that we don't use it anywhere. So, it's dropped in the latest PS. 

As for Runtime_AvailableLocaleOf(), locale ids come from the ICU rather than from a user so that there will never be an issue.  As such, I can't make a test with a very long locale id that would lead to the condition I'm guarding against. There's an existing test for AvailableLocaleOf(). Passing it at least shows that I'm not breaking. 

The only remaining case (canonicalize) is covered by a test I'm adding here.

Patch-set: 7
Reviewer: Gerrit User 1151758 <1151758@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>