Update patch set 4
Patch Set 4: Code-Review+1

Thanks for fixing this so fast!

This is a hard bug to tickle because of the constant folding we do for accessing WASM locals. Can it also be triggered with WASM memory accesses? If so, please add a test that triggers it. If not, my CL will definitely tickle it.

Patch-set: 4
Reviewer: Gerrit User 1119499 <1119499@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1