[cleanup] Introduce BuiltinsAssembler
This new CodeStubAssembler subclass is a shared superclass for
mksnapshot-time code generation (in particular, for Builtins and
Interpreter).
As a first step, it contains "BitwiseOp" and a templatized unification
of the "TaggedToWord32[OrBigInt][WithFeedback]" and
"TaggedToNumeric[WithFeedback]" helpers.

Over time, we can pull everything out of CSA into BA that is not
required for the shipping binary.

Bug: v8:6921
Change-Id: Iae5dcc4c50c7fde3423f801cb5484de337381ce6