Merged: Revert "Enable icu-timezone-data by default"
This reverts commit d9a25842d3aba82ead5e06c7fd7cb4896e6ad202.

Reason for revert:

I'm reverting this CL for a few reasons. #2 is the most significant and
I should have thought of that before making a switch. Sorry for that.

1) perf-regression: http://crbug.com/769706
2) http://crbug.com/612010 : ICU timezone update is not propagated to
zygote process so that new tabs will hold on to an old timezone even
after a timezone change on Linux and Chrome OS.

3)  http://crbug.com/754053  : OS timezone detection issues on macOS
10.13, Ubutu 16, RHEL 7, SuSe Linux 12 or newer. ; it's being fixed. So,
it actually ok.

4)  http://crbug.com/771868  : timezone wrong in gmail: If it's due to

Revision: 28ef8dc7009d

BUG=chromium:612010,chromium:769706,chromium:771868,v8:2137,v8:6031,v8:6076
LOG=N
NOTRY=true
NOPRESUBMIT=true
NOTREECHECKS=true
R=adamk@chromium.org, hablich@chromium.org

Change-Id: Id0093c2bc69e5bc1d3b7147b7bbcc633ed625a45