Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Wondering why this doesn't break in normal optimized debug builds?

This is highly nondeterministic. We pass a float or double around in a local variable, and depending on where the (C++) compiler decides to store that variable, the quiet NaN bit might flip (e.g. when it is pushed to the stack).
These bugs only show up in very specific builds. That's why it's important to test them more, as done in this CL ;)

Patch-set: 2
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>