[WIP] implement destructuring in BytecodeGenerator
This is a work in progress, and while it does work in some simple test
cases, it's quite likely that it's not there yet. Since the initial chunk
of this patch was written on MSVC, and debugging anything on MSVC is
horrible, I'm uploading this CL to make it easier to test this from
machines which can launch lldb.

But I suspect that it will be make certain things a bit better, slightly
reducing frame size when destructuring is used, etc, possibly shrinking
generated bytecode a bit.

Bug:
Change-Id: I888e5aeb5b22c10353fff7c971e88cf2b283b2a3