Update patch set 2
Patch Set 2:

Surprisingly, this revert seems to have fixed the arm64 gc stress failure in one of the wasm interpreter tests:
https://logs.chromium.org/v/?s=chromium%2Fbb%2Fclient.v8.ports%2FV8_Linux_-_arm64_-_sim_-_gc_stress%2F8935%2F%2B%2Frecipes%2Fsteps%2FMjsunit%2F0%2Flogs%2Finterpreter-mixed%2F0

Can this be related to this CL?
I was not able to reproduce the error, not even with the binary from the bots, so I have no further information what actually went wrong there. All I see is a nullptr access eventually.

Patch-set: 2
CC: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>