Update patch set 16
Patch Set 16:

> > I understand the current situation.  I understand that we'd like to brush up a way to mark non-node wrappers "active" in order to minimize the number of wrappers alive.  Will look when I have time.
> 
> To clarify (I think you meant "non-active" in your post): As far as I can understand, all non-node wrapper are currently already marked as active. The idea is to skip marking active some of them where we know that we can get rid of them.

Sorry for the confusion.  My understanding is that we're currently marking non-node wrappers "active" too aggressively, even when some of non-node wrappers do not need to be alive.  So, we'd like to improve this marking so that we'll skip marking for no-need-to-be-alive wrappers.  I think that my understanding is the same of yours.  Sorry for poor wording.

Patch-set: 16