Update patch set 3
Patch Set 3: Code-Review-1

I'm not trying to sound defensive, but is there a particular reason this needs to be in upstream V8 at this early stage? From what I read, the goal is to port Node.js to CloudABI. This apparently requires a series of changes that will depend on this flag. Why do you need to have this in upstream in V8 as opposed to maintain your own fork?

This aside, we are moving off GYP as build system. Does CloudABI support building with GN?

Patch-set: 3
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1