[log] Support first function execution logging with --log-function-events
This CL introduces a FirstFunctionExecution builtin which is used instead of
the default InterpreterEntryTramponline when --log-function-events is
active. The new builtin will do a runtime call for the logging and change
the JSFunction's code back to the usual InterpreterEntryTramponline before
resuming execution.

Using the new builtin tracing first function execution adds no overhead when
inactive.

Bug: chromium:757467
Change-Id: Icd11e9ffc139e03b50457b89a229e233d24c4915