Update patch set 16
Patch Set 16:

(2 comments)

High-level point - I'm not super keen on this approach as it complicates the logic as to whether a function is interpreted or not. 

There is also the issue that this could log the first execution of a SFI multiple times - i.e., if the IFET is on the SFI and two JSFunctions are created from that SFI before either is executed, then both JSFunctions will point to the IFET, and so both will log the first execution when they are actually called.

An alternative that might work and still avoid overhead when the flag is turned off - we always check the OptimizationMarker in the FeedbackVector when invoking a function in the IET. If we add a new marker (LogFirstRun) and have new feedback vectors are initialized with this marker if the flag is enabled (otherwise with OptimizationMarker::kNone as now). We would probably want to rename OptimizationMarker to be something like OnEntryMarker or similar. WDYT

+leszek for thoughts.

Patch-set: 16
Reviewer: Gerrit User 1175774 <1175774@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>