Update patch set 16
Patch Set 16:

> There is also the issue that this could log the first execution of a SFI multiple times - i.e., if the IFET is on the SFI and two JSFunctions are created from that SFI before either is executed, then both JSFunctions will point to the IFET, and so both will log the first execution when they are actually called.

Right, this isn't even hypothetical because it's the same reason we introduced the optimization marker in the first place. An example could be:

    var funcs = []
    for (var i = 0; i < 2; ++i) {
      funcs.push(/* Break IIFE heusistic --> */ (function() {
        ...
      }) );
    }
    for (var i = 0; i < 2; ++i) {
      funcs[i]();
    }

Similar stuff can happen with recursive calls.

> An alternative that might work and still avoid overhead when the flag is turned off - we always check the OptimizationMarker in the FeedbackVector when invoking a function in the IET. If we add a new marker (LogFirstRun) and have new feedback vectors are initialized with this marker if the flag is enabled (otherwise with OptimizationMarker::kNone as now). We would probably want to rename OptimizationMarker to be something like OnEntryMarker or similar.

I'm ok with this approach, though we'd have to do a reasonable amount of renaming to make it make sense, and we'd have to be careful with --always-opt. It'd also introduce a 1 branch overhead for functions that are currently in the optimization queue, probably not a big deal in the grand scheme of things.

Another (slightly horrible) alternative is having a log-first-run builtin be placed as a code pointer in the optimized code slot. This would, however, complicate IsOptimized checks, so there's not much advantage, I'm just listing it for completeness.

Patch-set: 16
Reviewer: Gerrit User 1175774 <1175774@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>